scripts/base/EaseSetRelativeCollision: Make script more slightly robust against Editor errors
For some reason, UpdateColliders would still inadvertently get set to true
(even though that never happens from the code).

Set another (dumb) field to ignore UpdateCollider for a frame, and force it
to false in the meanwhile.

This still throws an error if you Undo in the Editor, but I don't think I
want to bother trying to fix Unity's issues anymore...