scripts/level/InitialPlayerPosition: Skip checkpoints lower than the initial position
InitialPlayerPosition is assigned a "checkpoint" index (used by the game
when deciding at which position the player should be spawned) based on the
vertical position of the object. However, this breaks if a checkpoint is
lower than the starting position.

Automatically assign position 0 to the initial position and skip requesting
it a checkpoint index.