scripts/entity/InputControlled: Fix sending JustDropped on pulling from a ledge
If you pulled a block while hanging on the ledge, the game wouldn't trigger
the 'JustDropped' event on the block bellow. The main situation where this
would be noticed is on cracked blocks, since these wouldn't break/start
breaking in that specific situation.

Fix sending 'JustDropped' events when pulling a block while hanging on a
ledge.