scripts/util/Input: Reject and revert invalid configurations
Last release, a new action was added (dropping from the ledge). However,
that action wouldn't be accounted for in any currently saved configuration,
which would cause the action to simply never be triggered. A simple fix
from the user side is to revert the input to the defaults and remap it.

Check if the saved configuration is valid, and revert it if it's invalid.