scripts/util/Input: Invert d-pad up/down
Some people reported that d-pad up and down were reversed for them when
using a dual shock 4.

However, since some of the menu inputs are hard-coded, to try to prevent
people from locking themselves from operating the menus, simply inverting
the d-pad in the rebindable controls wouldn't be enough...

So invert the default values for d-pad up/down in map C and remove the
hard-coded values for d-pad controls in menus.